+++ mllc/src/ast.rs
@@ -101,7 +101,7 @@ pub enum Decl {
     },
 }
 
-#[derive(Debug, Clone, Copy)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum Assoc {
     Left,
     Right,